skill model modified !

+++ lib/features/user_profile/views/screens/add_edit_technical_skill_screen.dart
@@ -106,7 +106,7 @@ class _AddEditTechnicalSkillState extends State<AddEditTechnicalSkill> {
       if(correctInput(searchController.text)){
         var skillInfo = SkillInfo(
           profSkillId: widget.skillInfo?.profSkillId,
-          rating: int.parse(ratingController.text),
+          rating: double.parse(ratingController.text),
           skill: _selectedSkill,
         );
